+++ db/migrate/20210504085525_add_postgraduate_doctoral_loan_to_early_career_payments_eligibilities.rb
@@ -1,0 +1,5 @@
+class AddPostgraduateDoctoralLoanToEarlyCareerPaymentsEligibilities < ActiveRecord::Migration[6.0]
+  def change
+    add_column :early_career_payments_eligibilities, :postgraduate_doctoral_loan, :boolean
+  end
+end